Changed function for storing a data in APCu
According to this documentation http://php.net/manual/en/function.apcu-add.php it will return false when it fails to save a data. But it also means a case when the key already exsits in APCu and the data will not be overwritten. So better will be to use this method http://php.net/manual/en/function.apcu-store.php. This method overwrites data when the key exists, so false is returned only in case of real error.